Move mappings commands into a different package
