+++ src/main/java/com/mcmoddev/mmdbot/modules/commands/CommandModule.java
@@ -20,8 +20,8 @@ import com.mcmoddev.mmdbot.modules.commands.general.info.CmdMinecraftVersion;
 import com.mcmoddev.mmdbot.modules.commands.general.info.CmdPaste;
 import com.mcmoddev.mmdbot.modules.commands.general.info.CmdSearch;
 import com.mcmoddev.mmdbot.modules.commands.general.info.CmdXy;
-import com.mcmoddev.mmdbot.modules.commands.info.CmdMappings;
-import com.mcmoddev.mmdbot.modules.commands.info.CmdTranslateMappings;
+import com.mcmoddev.mmdbot.modules.commands.general.mappings.CmdMappings;
+import com.mcmoddev.mmdbot.modules.commands.general.mappings.CmdTranslateMappings;
 import com.mcmoddev.mmdbot.modules.commands.server.CmdGuild;
 import com.mcmoddev.mmdbot.modules.commands.server.CmdReadme;
 import com.mcmoddev.mmdbot.modules.commands.server.CmdRoles;
